+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Custom dimensions to `trackPage` for Google Analytics
+- Tombstone for withdrawn preprints
 
 ### Fixed
 - Use `name` field for contributors on preprint-detail page