+++ CHANGELOG.md
@@ -4,9 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.12.0] - 2021-04-27
 ### Added
 - pin node@8.17.0 and yarn@1.21.1 with volta
+
+### Changed
+- show all authors for moderator preprint detail page
 
 ## [0.11.0] - 2019-07-30
 ### Changed